+++ src/Ext/Rules/Body.cpp
@@ -375,6 +375,24 @@ void RulesExt::ExtData::LoadBeforeTypeData(RulesClass* pThis, CCINIClass* pINI)
 
 		this->AIScriptsLists.emplace_back(std::move(objectsList));
 	}
+
+	// Section AITriggersList
+	int triggerItemsCount = pINI->GetKeyCount("AITriggersList");
+	for (int i = 0; i < triggerItemsCount; ++i)
+	{
+		std::vector<AITriggerTypeClass*> objectsList;
+
+		char* context = nullptr;
+		pINI->ReadString("AITriggersList", pINI->GetKeyName("AITriggersList", i), "", Phobos::readBuffer);
+
+		for (char *cur = strtok_s(Phobos::readBuffer, Phobos::readDelims, &context); cur; cur = strtok_s(nullptr, Phobos::readDelims, &context))
+		{
+			AITriggerTypeClass* pNewTrigger = GameCreate<AITriggerTypeClass>(cur); // Note: Don't use ::FindOrAllocate(cur) here...
+			objectsList.emplace_back(pNewTrigger);
+		}
+
+		this->AITriggersLists.emplace_back(std::move(objectsList));
+	}
 }
 
 // this should load everything that TypeData is not dependant on
@@ -411,6 +429,7 @@ void RulesExt::ExtData::Serialize(T& Stm)
 	Stm
 		.Process(this->AITargetTypesLists)
 		.Process(this->AIScriptsLists)
+		.Process(this->AITriggersLists)
 		.Process(this->Storage_TiberiumIndex)
 		.Process(this->HarvesterDumpAmount)
 		.Process(this->InfantryGainSelfHealCap)