+++ src/Ext/Rules/Body.h
@@ -2,6 +2,9 @@
 
 #include <CCINIClass.h>
 #include <RulesClass.h>
+#include <GameStrings.h>
+#include <AITriggerTypeClass.h>
+
 #include <Utilities/Container.h>
 #include <Utilities/Constructs.h>
 #include <Utilities/Template.h>
@@ -31,6 +34,7 @@ public:
 	public:
 		std::vector<std::vector<TechnoTypeClass*>> AITargetTypesLists;
 		std::vector<std::vector<ScriptTypeClass*>> AIScriptsLists;
+		std::vector<std::vector<AITriggerTypeClass*>> AITriggersLists;
 
 		Valueable<int> Storage_TiberiumIndex;
 		Valueable<float> HarvesterDumpAmount;