+++ src/Ext/Team/Body.cpp
@@ -22,6 +22,8 @@ void TeamExt::ExtData::Serialize(T& Stm)
 		.Process(this->ForceJump_RepeatMode)
 		.Process(this->TeamLeader)
 		.Process(this->PreviousScriptList)
+		.Process(this->TriggersSideIdx)
+		.Process(this->TriggersHouseIdx)
 		;
 }